refactor: rewrite auth context to zustand store
